package import change and rename banner dismiss dialog for payments screen for better readability
