+++ WooCommerce/src/main/kotlin/com/woocommerce/android/ui/compose/component/banner/Banner.kt
@@ -1,4 +1,4 @@
-package com.woocommerce.android.ui.compose.component
+package com.woocommerce.android.ui.compose.component.banner
 
 import android.content.res.Configuration
 import androidx.compose.foundation.Image
