+++ WooCommerce/src/main/kotlin/com/woocommerce/android/ui/orders/list/OrderListFragment.kt
@@ -41,8 +41,8 @@ import com.woocommerce.android.model.FeatureFeedbackSettings.FeedbackState
 import com.woocommerce.android.tools.SelectedSite
 import com.woocommerce.android.ui.base.TopLevelFragment
 import com.woocommerce.android.ui.base.UIMessageResolver
-import com.woocommerce.android.ui.compose.component.OrderListBannerDismissDialog
-import com.woocommerce.android.ui.compose.component.OrderListScreenBanner
+import com.woocommerce.android.ui.compose.component.banner.OrderListBannerDismissDialog
+import com.woocommerce.android.ui.compose.component.banner.OrderListScreenBanner
 import com.woocommerce.android.ui.compose.theme.WooThemeWithBackground
 import com.woocommerce.android.ui.feedback.SurveyType
 import com.woocommerce.android.ui.main.MainActivity
